% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAN640SumOfSquares-package.R
\docType{package}
\name{BAN640SumOfSquares}
\alias{BAN640SumOfSquares}
\alias{BAN640}
\alias{BAN640SumOfSquares-package}
\title{BAN 640 Sum of Squares Assignment}
\description{
Simple package used to demonstrate how to make a package by
implementing the sum of squares function.
}
\author{
Mark Gonzalez
}
\keyword{BAN640}
\keyword{Gonzalez}
\keyword{Squares}
