% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squares_data.R
\docType{data}
\name{squares_data}
\alias{squares_data}
\title{This is data to be included in my package}
\format{
A data frame with the first 10 sum of squares for natural numbers
\itemize{
\item{x}{Sequence 1:10 represents n in sum_of_squares function.}
\item{y}{Sum of Squares based on x value.}
}
}
\description{
This is data to be included in my package
}
\author{
Mark Gonzalez \email{markgo1@mail.regent.edu}
}
