% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_of_squares.R
\name{sum_of_squares}
\alias{sum_of_squares}
\title{Sum of Square Calculation}
\usage{
sum_of_squares(n)
}
\arguments{
\item{n}{The first n Natural Numbers}
}
\value{
The Sum of Squares
}
\description{
Calculate the Sum of Squares for the First n Natural Numbers
}
\examples{
sos1 <- sum_of_squares(50);
}
